backend/agent.py
from google.adk import Agent, Model

# Define the Comedian Agent
joker_agent = Agent(
    model=Model(model_name="gemini-1.5-flash"), # Use a fast model for responsiveness
    system_instruction="""
    You are a professional stand-up comedian. 
    Your goal is to tell funny, engaging, and sometimes sarcastic jokes based on the user's topic.
    
    Guidelines:
    - Keep it punchy and concise (unless asked for a story).
    - Use emojis to enhance the delivery 🎤.
    - If the topic is boring, make fun of the topic itself before telling the joke.
    - Avoid offensive or inappropriate content, but playful roasting is allowed.
    """,
)

# This is for the ADK API Server to pick up
agent = joker_agent
